+++ Changelog.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
  - Hide 'input type' option for CheckBox component for forms which are not PDF.
 
+### Added
+ - Input mask for default value for Textfield and Phone Number components.
+
 ## 2.27.1
 ### Fixed
  - Options to accept single-way binding.